requirements.txt
# DeepSeek-OCR specific dependencies (NEW)
einops
addict
easydict
transformers==4.46.3  # Bumped from 4.45.0 for DeepSeek support
tokenizers>=0.20.3

# Core dependencies
torch>=2.0.0
pillow>=10.0.0
pdf2image>=1.16.3
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=0.27.0
qwen-vl-utils
sentencepiece>=0.1.99
protobuf>=3.20.0
python-dotenv>=1.0.0

# API and UI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
python-multipart>=0.0.6

# Optional: Flash Attention 2 for faster inference (requires CUDA)
# Note: This often fails to install via requirements.txt. 
# It is better to install it manually: pip install flash-attn --no-build-isolation
# flash-attn>=2.7.3

# System dependencies (install via apt-get/brew):
# - poppler-utils (required for pdf2image)